tieqep: add patch to fix unhandled fault on 4.9 kernel
This patch adds a fix for the tieqep driver to avoid unhandled fault
when reading or writing to memory mapped eQEP registers.

The fix is to call pm_runtime_get_sync() at the beginning of any functions
that will read from or write to eQEP registers.  This should ensure that
the eQEP peripheral is running and its clock is enabled.

Before this patch, an attempt to read the position file via sysfs would
results in a segmentation fault.  The kernel log would contain this error:

[ 2591.653471] Unhandled fault: external abort on non-linefetch (0x1028) at 0xfa304180
[ 2591.915165] [<bf005310>] (eqep_get_position [tieqep]) from [<c08930d0>] (dev_attr_show+0x2c/0x58)

Details:
https://gist.github.com/pdp7/fe07082d23f2bfbc362c733a7b0aea72

BeagleBoard mailing list thread:
https://groups.google.com/d/msg/beagleboard/_TdTH7oPEXE/MNvU-mY6DgAJ